requirements.txt
influxdb-client
pyzmq
requests==2.31.0
flask

